refactor: manage logic outside of the TicketCard component
