--- src/Components/Workflow/components/WorkflowColumn/WorkflowColumn.css
@@ -1,3 +1,0 @@
-.colone-ticket {
-  flex: 0 0 340px;
-}
