--- src/Components/Workflow/components/WorkflowColumn/index.js
@@ -1,1 +1,0 @@
-export * from "./WorkflowColumn"
